tts_rating_app/requirements.txt
Flask==3.1.2
Authlib==1.6.5
requests==2.32.5
gunicorn==23.0.0
python-dotenv==1.2.1
